Create DbAppConfig
