+++ db-commons/src/main/resources/reference.conf
@@ -2,7 +2,7 @@ bitcoin-s {
 
   network = regtest # regtest, testnet3, mainnet, signet
   datadir = ${HOME}".bitcoin-s/"
-  sqlite = {
+  dbDefault = {
     dataSourceClass = slick.jdbc.DatabaseUrlDataSource
     profile = "slick.jdbc.SQLiteProfile$"
 
@@ -25,7 +25,7 @@ bitcoin-s {
     }
   }
 
-  node = ${bitcoin-s.sqlite}
+  node = ${bitcoin-s.dbDefault}
 
   node {
     peers = ["localhost"] # a list of peer addresses in form "hostname:portnumber"
@@ -48,7 +48,7 @@ bitcoin-s {
   }
 
 
-  chain = ${bitcoin-s.sqlite}
+  chain = ${bitcoin-s.dbDefault}
   chain {
     neutrino {
       filter-header-batch-size.default = 2000
@@ -74,7 +74,7 @@ bitcoin-s {
     # }
   }
 
-  wallet = ${bitcoin-s.sqlite}
+  wallet = ${bitcoin-s.dbDefault}
     # settings for wallet module
   wallet {
       defaultAccountType = legacy # legacy, segwit, nested-segwit
@@ -113,7 +113,7 @@ bitcoin-s {
         rpcport = 9999
     }
 
-  oracle = ${bitcoin-s.sqlite}
+  oracle = ${bitcoin-s.dbDefault}
   oracle {
     # this config key is read by Slick
     db {
@@ -129,7 +129,7 @@ bitcoin-s {
     # }
   }
 
-  test = ${bitcoin-s.sqlite}
+  test = ${bitcoin-s.dbDefault}
   test {
     # this config key is read by Slick
     db {
