+++ dlc-oracle/src/main/scala/org/bitcoins/dlc/oracle/config/DLCOracleAppConfig.scala
@@ -1,7 +1,5 @@
 package org.bitcoins.dlc.oracle.config
 
-import java.nio.file.{Files, Path}
-
 import com.typesafe.config.Config
 import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.crypto.ExtKeyVersion.SegWitMainNetPriv
@@ -18,12 +16,13 @@ import org.bitcoins.keymanager.WalletStorage
 import org.bitcoins.keymanager.bip39.BIP39KeyManager
 import org.bitcoins.keymanager.config.KeyManagerAppConfig
 
+import java.nio.file.{Files, Path}
 import scala.concurrent.{ExecutionContext, Future}
 
 case class DLCOracleAppConfig(
     private val directory: Path,
     private val confs: Config*)(implicit val ec: ExecutionContext)
-    extends AppConfig
+    extends DbAppConfig
     with DbManagement
     with JdbcProfileComponent[DLCOracleAppConfig] {
