+++ testkit/src/main/resources/reference.conf
@@ -2,7 +2,7 @@ bitcoin-s {
     datadir = ${HOME}/.bitcoin-s
     network = regtest # regtest, testnet3, mainnet, signet
 
-    sqlite = {
+    dbDefault = {
         dataSourceClass = slick.jdbc.DatabaseUrlDataSource
         profile = "slick.jdbc.SQLiteProfile$"
 
@@ -25,7 +25,7 @@ bitcoin-s {
         }
     }
 
-    oracle = ${bitcoin-s.sqlite}
+    oracle = ${bitcoin-s.dbDefault}
     oracle {
         # this config key is read by Slick
         db {
@@ -42,7 +42,7 @@ bitcoin-s {
     }
 
 
-    node = ${bitcoin-s.sqlite}
+    node = ${bitcoin-s.dbDefault}
     node {
         mode = neutrino # neutrino, spv
 
@@ -65,7 +65,7 @@ bitcoin-s {
 
 
 
-    chain = ${bitcoin-s.sqlite}
+    chain = ${bitcoin-s.dbDefault}
     chain {
         neutrino {
             filter-header-batch-size.default = 2000
@@ -93,7 +93,7 @@ bitcoin-s {
     }
 
 
-    wallet = ${bitcoin-s.sqlite}
+    wallet = ${bitcoin-s.dbDefault}
     # settings for wallet module
     wallet {
         defaultAccountType = legacy # legacy, segwit, nested-segwit
@@ -134,7 +134,7 @@ bitcoin-s {
         rpcport = 9999
     }
 
-    test = ${bitcoin-s.sqlite}
+    test = ${bitcoin-s.dbDefault}
     test {
         # this config key is read by Slick
         db {
